+++ lib/ncore/methods/create.rb
@@ -3,28 +3,29 @@ module NCore
     extend ActiveSupport::Concern
 
     module ClassMethods
-      def create!(attribs={}, api_creds=nil)
-        obj = create(attribs, api_creds)
-        if obj.errors.any?
+      def create!(attribs={})
+        obj = create(attribs)
+        if obj.errors?
           raise parent::RecordInvalid, obj
         end
         obj
       end
 
       # always returns a new object; check .errors? or .valid? to see how it went
-      def create(attribs={}, api_creds=nil)
-        params = {json_root => attribs}
-        parsed, creds = request(:post, url, api_creds, params)
-        new(attribs, creds).send(:load, parsed)
+      def create(attribs={})
+        params = parse_request_params(attribs)
+        obj = new({}, params[:credentials])
+        obj.send :create, params
+        obj
       end
     end
 
     private
 
     def create(attribs={})
-      params = {json_root => attribs}
-      parsed, @api_creds = request(:post, self.class.url, api_creds, params)
-      load(data: attribs) if parsed[:errors].any?
+      params = parse_request_params(attribs, json_root: json_root).reverse_merge credentials: api_creds
+      parsed, @api_creds = request(:post, self.class.url, params)
+      load(data: params[json_root]) if parsed[:errors].any?
       load(parsed)
       errors.empty? ? self : false
     end