+++ lib/ncore/methods/find.rb
@@ -3,13 +3,15 @@ module NCore
     extend ActiveSupport::Concern
 
     module ClassMethods
-      def find(id, params={}, api_creds=nil)
-        o = new({id: id}, api_creds)
+      def find(id, params={})
+        raise(parent::RecordNotFound, "Cannot find id=nil") if id.blank?
+        params = parse_request_params(params)
+        o = new({id: id}, params[:credentials])
         o.reload(params)
       end
 
-      def retrieve(id, params={}, api_creds=nil)
-        find id, params, api_creds
+      def retrieve(id, params={})
+        find id, params
       rescue parent::RecordNotFound
         false
       end
@@ -17,7 +19,8 @@ module NCore
 
     def reload(find_params={})
       return if id.blank?
-      parsed, @api_creds = request(:get, url, api_creds, find_params)
+      params = parse_request_params(find_params).reverse_merge credentials: api_creds
+      parsed, @api_creds = request(:get, url, params)
       @attribs = {}.with_indifferent_access
       load(parsed)
     end