+++ lib/ncore/methods/update.rb
@@ -3,26 +3,28 @@ module NCore
     extend ActiveSupport::Concern
 
     module ClassMethods
-      def update!(id, attribs, api_creds=nil)
-        obj = new({id: id}, api_creds)
-        obj.save!(attribs)
+      def update!(id, attribs)
+        obj = update(id, attribs)
+        if obj.errors?
+          raise parent::RecordInvalid, obj
+        end
+        obj
       end
 
       # always returns a new object; check .errors? or .valid? to see how it went
-      def update(id, attribs, api_creds=nil)
-        obj = new({id: id}, api_creds)
-        obj.save(attribs)
+      def update(id, attribs)
+        raise(parent::RecordNotFound, "Cannot update id=nil") if id.blank?
+        params = parse_request_params(attribs)
+        obj = new({id: id}, params[:credentials])
+        obj.update params
         obj
       end
     end
 
-
-    private
-
     def update(attribs={})
-      params = {json_root => attribs}
-      parsed, @api_creds = request(:put, url, api_creds, params)
-      load(data: attribs) if parsed[:errors].any?
+      params = parse_request_params(attribs, json_root: json_root).reverse_merge credentials: api_creds
+      parsed, @api_creds = request(:put, url, params)
+      load(data: params[json_root]) if parsed[:errors].any?
       load(parsed)
       errors.empty? ? self : false
     end